refactor(core): move unfurl metadata off route helpers
Scrape service no longer imports the chat-room routes helper.
